+++ FP/GameObjects.h
@@ -7,15 +7,15 @@
 
 struct Vect_2D {
 
-	float x;
-	float y;
+	int x;
+	int y;
 };
 
 struct Circle {
 
-	float x;
-	float y;
-	float r;
+	int x;
+	int y;
+	int r;
 };
 
 template <typename T>
@@ -24,10 +24,10 @@ class GameObjects{
 protected:
 
 	Vect_2D _velocity;
-	float _speed;
+	int _speed;
 
-	float _curX;
-	float _curY;
+	int _curX;
+	int _curY;
 	SDL_Texture* _texture;	// TODO need to destroy textures // SDL_DestroyTexture(t);
 	T _boundingBox;
 
@@ -35,17 +35,17 @@ public:
 
 	// Accessors
 	Vect_2D velocity() { return _velocity; }
-	float curX() { return _curX; }
-	float curY() { return _curY; }
+	int curX() { return _curX; }
+	int curY() { return _curY; }
 	T boundingBox() { return _boundingBox;  }
 	SDL_Texture* getTexture() { return _texture;  }
 
 	// Modifiers
 	void setTexture(SDL_Texture* t) { _texture = t; }
 	void setVelocity(Vect_2D v) { _velocity = v;  }
-	void setSpeed(float s) { _speed = s;  }
-	void curX(float n); 
-	void curY(float n); 
+	void setSpeed(int s) { _speed = s;  }
+	void curX(int n);
+	void curY(int n);
 	
 	// Special functions
 	virtual void render(SDL_Renderer* renderer) = 0;
@@ -61,7 +61,7 @@ private:
 public:
 
 	Ball() = default;
-	Ball(const float& x, const float& y, const float& r);
+	Ball(const int& x, const int& y, const int& r);
 
 	// Functions
 	void render(SDL_Renderer* renderer) override;
@@ -73,7 +73,7 @@ class Platform : public GameObjects<SDL_Rect> {
 public:
 
 	Platform() = default;
-	Platform(const float& x, const float& y, const float& w, const float& h);
+	Platform(const int& x, const int& y, const int& w, const int& h);
 
 	// Functions
 	void moveUp();
@@ -90,13 +90,13 @@ void GameObjects<T>::move() {
 }
 
 template <typename T>
-void GameObjects<T>::curX(float n) {
+void GameObjects<T>::curX(int n) {
 	_curX = n;
 	updateBoundingBox();
 }
 
 template <typename T>
-void GameObjects<T>::curY(float n) {
+void GameObjects<T>::curY(int n) {
 	_curY = n;
 	updateBoundingBox();
 }
